+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- LocalData: add optional entity_short_name attribute to local data configuration to specify the entity the local data
+attribute is defined for. It can be omitted if the local data attribute is defined for all entities or there is only one entity using
+local data attributes.
 - AbstractLocalDataEventSubscriber: allow overriding the getting of the local data attribute value, e.g. for attributes
 that can be provided by a data connector, but whose values are not included in the source system's standard set of attributes.